extract clone cell.
