+++ src/vs/workbench/contrib/notebook/common/model/notebookCellTextModel.ts
@@ -7,6 +7,7 @@ import { Emitter, Event } from 'vs/base/common/event';
 import { ICell, NotebookCellOutputsSplice, CellKind, NotebookCellMetadata, NotebookDocumentMetadata, TransientOptions, IOutputDto, ICellOutput } from 'vs/workbench/contrib/notebook/common/notebookCommon';
 import { PieceTreeTextBufferBuilder } from 'vs/editor/common/model/pieceTreeTextBuffer/pieceTreeTextBufferBuilder';
 import { URI } from 'vs/base/common/uri';
+import * as UUID from 'vs/base/common/uuid';
 import * as model from 'vs/editor/common/model';
 import { Range } from 'vs/editor/common/core/range';
 import { Disposable } from 'vs/base/common/lifecycle';
@@ -178,4 +179,28 @@ export class NotebookCellTextModel extends Disposable implements ICell {
 		this._textBuffer = emptyDisposedTextBuffer;
 		super.dispose();
 	}
+}
+
+export function cloneMetadata(cell: NotebookCellTextModel) {
+	return {
+		editable: cell.metadata?.editable,
+		breakpointMargin: cell.metadata?.breakpointMargin,
+		hasExecutionOrder: cell.metadata?.hasExecutionOrder,
+		inputCollapsed: cell.metadata?.inputCollapsed,
+		outputCollapsed: cell.metadata?.outputCollapsed,
+		custom: cell.metadata?.custom
+	};
+}
+
+export function cloneNotebookCellTextModel(cell: NotebookCellTextModel) {
+	return {
+		source: cell.getValue(),
+		language: cell.language,
+		cellKind: cell.cellKind,
+		outputs: cell.outputs.map(output => ({
+			outputs: output.outputs,
+			/* paste should generate new outputId */ outputId: UUID.generateUuid()
+		})),
+		metadata: cloneMetadata(cell)
+	};
 }
